Refactor Widget/Window types and includes
